WIP -- hooks: Add dkimpy support
DO NOT SUBMIT - THIS HAS ISSUES

This patch adds support in the default hook for using dkimpy for DKIM
signing.

FIXME: This can't be currently tested properly because dkimpy's
dkimverify can't be made to use custom DNS, or override the TXT values
in any way, so we can't verify that the generated signature is
reasonable.